validate first e2e test
